pyproject.toml
[project]
name = "pineneedle"
version = "0.1.0"
description = "A smart Python application that leverages LLMs to intelligently manage and customize resumes"
requires-python = ">=3.12"

dependencies = [
    "click",
    "markdown",
    "openai",
    "anthropic",
    "tomli-w",
    "pydantic-ai>=0.4.5",
    "python-dotenv",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "weasyprint",
    "questionary>=2.0.0",
]

[project.scripts]
pineneedle = "pineneedle.cli.main:main"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pineneedle*"] 
